apps/backend/pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Eli", email = "53414320+Eli-Chandler@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.8.6",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "bcrypt>=4.3.0",
    "browser-use>=0.5.5",
    "fastapi>=0.116.1",
    "mypy>=1.11",
    "py2pdf>=0.1.3",
    "pydantic[email]>=2.5.3",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.41",
]

[project.scripts]
job-agent = "job_agent:main"
dev = "api.main:start_dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.5",
    "ruff-lsp>=0.0.62",
]

